Move new pacts/mills types out of legacy files
Legacy files are now only fixed up or shrunk, never extended. The new
hexagonal additions land in their proper subdomain modules:

- pacts/services.py (NEW): cross-cutting service infrastructure —
  TransactionProtocol and the navigation tree (ServicesProtocol,
  ChronologyServicesProtocol, ChronologyPanelServicesProtocol).
- pacts/chronology.py: gains the CFP bounded-context types
  (CFPPersonalDataFieldServiceProtocol plus the form-context DTOs).
  Module docstring broadened from timetable-only to chronology subdomain.
- mills/chronology.py: gains CFPPersonalDataFieldService.
- pacts/legacy.py: strips the seven types it had grown; RootRequestProtocol
  now references ServicesProtocol via TYPE_CHECKING + forward ref to
  break the otherwise circular dep.
- mills/legacy.py: strips CFPPersonalDataFieldService and the imports it
  needed; pacts side cleaned up symmetrically.

Import sites updated: inits/services/chronology/{panel,root}.py and
the test module reach the symbols at their new homes; views/base.py
imports ServicesProtocol via TYPE_CHECKING.

All 1407 tests pass; lint, mypy, and importlinter clean.
