+++ src/ludamus/mills/chronology.py
@@ -1,4 +1,9 @@
-"""Timetable business logic for the agenda scheduling feature."""
+"""Chronology subdomain business logic.
+
+Currently spans the Timetable (agenda scheduling) and CFP (personal-data
+field management) bounded contexts. Split per `plans/hex_refactor.md` if
+the file grows past ~12 top-level members or 1000 lines.
+"""
 
 import math
 from collections import defaultdict
@@ -6,6 +11,7 @@ from datetime import date, datetime, timedelta, tzinfo
 from typing import TYPE_CHECKING
 
 from ludamus.pacts import (
+    FieldUsageSummary,
     NotFoundError,
     ScheduleChangeAction,
     ScheduleChangeLogData,
@@ -23,6 +29,8 @@ from ludamus.pacts.chronology import (
     HeatmapDayDTO,
     HeatmapDTO,
     HeatmapRowDTO,
+    PersonalDataFieldEditContextDTO,
+    PersonalDataFieldFormContextDTO,
     PreferredSlotRangeDTO,
     PreferredSlotViolationDTO,
     SessionPositionDTO,
@@ -37,10 +45,16 @@ if TYPE_CHECKING:
     from ludamus.pacts import (
         AgendaItemDTO,
         AreaDTO,
+        PersonalDataFieldCreateData,
+        PersonalDataFieldDTO,
+        PersonalDataFieldRepositoryProtocol,
+        PersonalDataFieldUpdateData,
+        ProposalCategoryRepositoryProtocol,
         SpaceDTO,
         TimeSlotDTO,
         UnitOfWorkProtocol,
     )
+    from ludamus.pacts.services import TransactionProtocol
 
 
 def _slot_windows_by_local_date(
@@ -645,3 +659,84 @@ class TimetableOverviewService:
                 )
             )
         return result
+
+
+class CFPPersonalDataFieldService:
+    """Backoffice operations for an event's personal-data fields."""
+
+    def __init__(
+        self,
+        transaction: TransactionProtocol,
+        fields: PersonalDataFieldRepositoryProtocol,
+        categories: ProposalCategoryRepositoryProtocol,
+    ) -> None:
+        self._transaction = transaction
+        self._fields = fields
+        self._categories = categories
+
+    def list_summaries(self, event_pk: int) -> list[FieldUsageSummary]:
+        fields = self._fields.list_by_event(event_pk)
+        usage_counts = self._fields.get_usage_counts(event_pk)
+        return [
+            FieldUsageSummary(
+                field=f,
+                required_count=usage_counts.get(f.pk, {}).get("required", 0),
+                optional_count=usage_counts.get(f.pk, {}).get("optional", 0),
+            )
+            for f in fields
+        ]
+
+    def get_create_form_context(self, event_pk: int) -> PersonalDataFieldFormContextDTO:
+        return PersonalDataFieldFormContextDTO(
+            categories=self._categories.list_by_event(event_pk)
+        )
+
+    def get_edit_form_context(
+        self, event_pk: int, field_slug: str
+    ) -> PersonalDataFieldEditContextDTO:
+        field = self._fields.read_by_slug(event_pk, field_slug)
+        categories = self._categories.list_by_event(event_pk)
+        field_cats = self._categories.get_personal_field_categories(field.pk)
+        return PersonalDataFieldEditContextDTO(
+            field=field,
+            categories=categories,
+            required_category_pks={pk for pk, req in field_cats.items() if req},
+            optional_category_pks={pk for pk, req in field_cats.items() if not req},
+        )
+
+    def create(
+        self,
+        event_pk: int,
+        data: PersonalDataFieldCreateData,
+        category_requirements: dict[int, bool],
+    ) -> PersonalDataFieldDTO:
+        with self._transaction.atomic():
+            field = self._fields.create(event_pk, data)
+            if category_requirements:
+                self._categories.add_field_to_categories(
+                    field.pk, category_requirements
+                )
+        return field
+
+    def update(
+        self,
+        event_pk: int,
+        field_slug: str,
+        data: PersonalDataFieldUpdateData,
+        category_requirements: dict[int, bool],
+    ) -> None:
+        field = self._fields.read_by_slug(event_pk, field_slug)
+        with self._transaction.atomic():
+            self._fields.update(field.pk, data)
+            self._categories.set_personal_field_categories(
+                field.pk, category_requirements
+            )
+
+    def delete(self, event_pk: int, field_slug: str) -> bool:
+        # Returns False when the field is in use by session types.
+        # NotFoundError on bad slug surfaces to the caller for distinct messaging.
+        field = self._fields.read_by_slug(event_pk, field_slug)
+        if self._fields.has_requirements(field.pk):
+            return False
+        self._fields.delete(field.pk)
+        return True
