+++ src/ludamus/mills/legacy.py
@@ -25,20 +25,12 @@ from ludamus.pacts import (
     FacilitatorData,
     FacilitatorDTO,
     FacilitatorMergeError,
-    FieldUsageSummary,
     HostPersonalDataEntry,
     MembershipAPIError,
     NotFoundError,
     PanelStatsDTO,
-    PersonalDataFieldCreateData,
-    PersonalDataFieldDTO,
-    PersonalDataFieldEditContextDTO,
-    PersonalDataFieldFormContextDTO,
-    PersonalDataFieldRepositoryProtocol,
-    PersonalDataFieldUpdateData,
     PersonalFieldRequirementDTO,
     ProposalCategoryDTO,
-    ProposalCategoryRepositoryProtocol,
     ProposeSessionResult,
     RequestContext,
     SessionData,
@@ -50,7 +42,6 @@ from ludamus.pacts import (
     TicketAPIProtocol,
     TimeSlotRequirementDTO,
     TrackDTO,
-    TransactionProtocol,
     UnitOfWorkProtocol,
     UserData,
     UserDTO,
@@ -532,87 +523,6 @@ class AcceptProposalService:
                     end_time=time_slot.end_time,
                 )
             )
-
-
-class CFPPersonalDataFieldService:
-    """Backoffice operations for an event's personal-data fields."""
-
-    def __init__(
-        self,
-        transaction: TransactionProtocol,
-        fields: PersonalDataFieldRepositoryProtocol,
-        categories: ProposalCategoryRepositoryProtocol,
-    ) -> None:
-        self._transaction = transaction
-        self._fields = fields
-        self._categories = categories
-
-    def list_summaries(self, event_pk: int) -> list[FieldUsageSummary]:
-        fields = self._fields.list_by_event(event_pk)
-        usage_counts = self._fields.get_usage_counts(event_pk)
-        return [
-            FieldUsageSummary(
-                field=f,
-                required_count=usage_counts.get(f.pk, {}).get("required", 0),
-                optional_count=usage_counts.get(f.pk, {}).get("optional", 0),
-            )
-            for f in fields
-        ]
-
-    def get_create_form_context(self, event_pk: int) -> PersonalDataFieldFormContextDTO:
-        return PersonalDataFieldFormContextDTO(
-            categories=self._categories.list_by_event(event_pk)
-        )
-
-    def get_edit_form_context(
-        self, event_pk: int, field_slug: str
-    ) -> PersonalDataFieldEditContextDTO:
-        field = self._fields.read_by_slug(event_pk, field_slug)
-        categories = self._categories.list_by_event(event_pk)
-        field_cats = self._categories.get_personal_field_categories(field.pk)
-        return PersonalDataFieldEditContextDTO(
-            field=field,
-            categories=categories,
-            required_category_pks={pk for pk, req in field_cats.items() if req},
-            optional_category_pks={pk for pk, req in field_cats.items() if not req},
-        )
-
-    def create(
-        self,
-        event_pk: int,
-        data: PersonalDataFieldCreateData,
-        category_requirements: dict[int, bool],
-    ) -> PersonalDataFieldDTO:
-        with self._transaction.atomic():
-            field = self._fields.create(event_pk, data)
-            if category_requirements:
-                self._categories.add_field_to_categories(
-                    field.pk, category_requirements
-                )
-        return field
-
-    def update(
-        self,
-        event_pk: int,
-        field_slug: str,
-        data: PersonalDataFieldUpdateData,
-        category_requirements: dict[int, bool],
-    ) -> None:
-        field = self._fields.read_by_slug(event_pk, field_slug)
-        with self._transaction.atomic():
-            self._fields.update(field.pk, data)
-            self._categories.set_personal_field_categories(
-                field.pk, category_requirements
-            )
-
-    def delete(self, event_pk: int, field_slug: str) -> bool:
-        # Returns False when the field is in use by session types.
-        # NotFoundError on bad slug surfaces to the caller for distinct messaging.
-        field = self._fields.read_by_slug(event_pk, field_slug)
-        if self._fields.has_requirements(field.pk):
-            return False
-        self._fields.delete(field.pk)
-        return True
 
 
 class PanelService:
