+++ src/ludamus/pacts/chronology.py
@@ -1,11 +1,26 @@
-"""Timetable DTOs and constants for the agenda scheduling feature."""
+"""Chronology subdomain DTOs and protocols.
 
+Currently spans the Timetable (agenda scheduling) and CFP (personal-data
+field management) bounded contexts. Split per `plans/hex_refactor.md` if
+the file grows past ~12 top-level members or 1000 lines.
+"""
+
+from dataclasses import dataclass
 from datetime import date, datetime
 from enum import StrEnum, auto
+from typing import Protocol
 
 from pydantic import BaseModel
 
-from ludamus.pacts.legacy import AgendaItemDTO, SpaceDTO
+from ludamus.pacts.legacy import (
+    AgendaItemDTO,
+    FieldUsageSummary,
+    PersonalDataFieldCreateData,
+    PersonalDataFieldDTO,
+    PersonalDataFieldUpdateData,
+    ProposalCategoryDTO,
+    SpaceDTO,
+)
 
 TIMETABLE_ROOM_PAGE_SIZE = 5
 TIMETABLE_SLOT_MINUTES = 60
@@ -129,3 +144,47 @@ class TrackProgressDTO(BaseModel):
     accepted_count: int
     scheduled_count: int
     progress_pct: int
+
+
+# --- CFP (personal-data field management) ---
+
+
+@dataclass
+class PersonalDataFieldFormContextDTO:
+    """Read aggregate for the personal-data-field create form."""
+
+    categories: list[ProposalCategoryDTO]
+
+
+@dataclass
+class PersonalDataFieldEditContextDTO:
+    """Read aggregate for the personal-data-field edit form."""
+
+    field: PersonalDataFieldDTO
+    categories: list[ProposalCategoryDTO]
+    required_category_pks: set[int]
+    optional_category_pks: set[int]
+
+
+class CFPPersonalDataFieldServiceProtocol(Protocol):
+    def list_summaries(self, event_pk: int) -> list[FieldUsageSummary]: ...
+    def get_create_form_context(
+        self, event_pk: int
+    ) -> PersonalDataFieldFormContextDTO: ...
+    def get_edit_form_context(
+        self, event_pk: int, field_slug: str
+    ) -> PersonalDataFieldEditContextDTO: ...
+    def create(
+        self,
+        event_pk: int,
+        data: PersonalDataFieldCreateData,
+        category_requirements: dict[int, bool],
+    ) -> PersonalDataFieldDTO: ...
+    def update(
+        self,
+        event_pk: int,
+        field_slug: str,
+        data: PersonalDataFieldUpdateData,
+        category_requirements: dict[int, bool],
+    ) -> None: ...
+    def delete(self, event_pk: int, field_slug: str) -> bool: ...
